[🛠️ LITTLE_FIXED] : 예외처리 추가 & 파일명 변경
